+++ WoWsShipBuilder.UI/Utilities/ScalableWindow.cs
@@ -1,0 +1,17 @@
+﻿using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Platform;
+using Avalonia.ReactiveUI;
+
+namespace WoWsShipBuilder.UI.Utilities
+{
+    public class ScalableWindow : Window, IScalableWindow
+    {
+        public void ProcessResizing(Size newSize, PlatformResizeReason resizeReason) => HandleResized(newSize, resizeReason);
+    }
+
+    public class ScalableReactiveWindow<T> : ReactiveWindow<T>, IScalableWindow where T : class
+    {
+        public void ProcessResizing(Size newSize, PlatformResizeReason resizeReason) => HandleResized(newSize, resizeReason);
+    }
+}