+++ WoWsShipBuilder.UI/Views/StartingMenuWindow.axaml.cs
@@ -1,11 +1,13 @@
+using System;
 using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
+using WoWsShipBuilder.UI.Utilities;
 using WoWsShipBuilder.UI.ViewModels;
 
 namespace WoWsShipBuilder.UI.Views
 {
-    public partial class StartingMenuWindow : Window
+    public partial class StartingMenuWindow : ScalableWindow
     {
         public StartingMenuWindow()
         {
@@ -31,6 +33,12 @@ namespace WoWsShipBuilder.UI.Views
             var window = new TestWindow();
             window.DataContext = new TestWindowViewModel();
             window.Show();
+        }
+
+        protected override void OnOpened(EventArgs e)
+        {
+            base.OnOpened(e);
+            this.HandleAndCheckScaling();
         }
 
         private void LoadBuild(object? sender, Avalonia.Interactivity.RoutedEventArgs e)